trim down event metadata, provide uuid to filter worker only
Disabled the printing of the target (module path) and thread name in each event. Instead, spawned threads have their thread ID as a field in their root span. Removed worker ID because it was redundant with thread ID. The filter worker was using the worker ID as a partition key, so now the filter worker just creates its own uuid when it spawns.