simplify emptiness check
Checking the length of the queue and then checking the length of the
vector of popped items is redundant. If the queue is empty, then the
vector will be empty. It's therefore sufficient to attempt to pop from
the queue and sleep/check if the length of the vector is zero.