add option for docker compose project
The value passed to the --project option is used for the -p option in docker compose. The default value is "boom-throughput", ensuring that the containers and volumes started by this script are completely isolated from the normal "boom" compose project. This commit also introduces an --alert-db option for specifying which mongodb database to use (default "boom").