Summerschool-inspired bugfixes/improvements (#213)
* the `spatial::xmatch` method had an issue with its first aggregation
stage, where code was outdated with regard to the other stages and the
code used read the data coming back from MongoDB. Also, some of the
fields we read from mongodb and expect to be f64 could in fact be read
as i64 by the driver (types are auto inferred by compass when you ingest
a catalog from there, so you don't have consistant typing in a
collection and numbers without digits might be considered to be i64 and
thereafter read as such).

* prv_candidates entry showed some duplication. This happens because
`prv_candidates` from ZTF is just an array of previous alerts they have
generated and stored in their DB, so when we add new entries to it and
just used MongoDB's `addToSet` operator. One thing that prv_candidates
does not have is the latest/current alert that it came with (it has all
the alerts up to but the included the current alert), so we ingest it to
make sure the current alert isn't ever missing in those arrays
(something we didn't do in Kowalski, and resulted in confusing,
incomplete lightcurves). Issue is for some ungodly reason the
`candidate` from the alert once included in the `prv_candidates` of a
future alert is NOT the same as the `candidate` from the initial
alert!!! Some floats vary ever so slightly (not significantly but well
for a computer than these values aren't equal) and so the `addToSet`
sees these values as distinct. So instead, we now use a custom
operator/function to deduplicate based on `jd` which really is the
unique key we care about for deduplication. For the `fp_hists` things
are a little more complicated so we might want to revisit that logic
(basically newer is better, so we could sort the arrays descending to
have newer values for the same JDs first which means we select them over
older ones). It's not too important right now, avoiding any kind of
duplication is more important imo.

* the kafka producers and consumers really need some retries, which can
be needed due to kafka availability or networking issues. While
producing/consuming on an old machine, my kafka was struggling to stay
alive and that meant the code just miserably failed instead of waiting
some amount of time for it to get back to life. In the producing case I
opted for a max number of retries + some sleep, where for the consumer
case I opted for retries until successful (so we never fail in
production while waiting for a remote server to come back to life, which
could take any amount of time).

* when reading the output of filters we ran on alerts, we had some
failures which the current logging setup just kind of ignored. I mean we
still knew an error happened in a method but not the line. Here I added
the `inspect_error()` calls needed to see said errors properly. Filters
that would yield incorrect documents should not be loaded to begin with,
but that's solved by another PR we have open (#139).

* the kafka consumer binary wasn't using the specified number of
processes and always defaulting to 1.