Throughput reports bugfix (#231)
Looks like we weren't able to read the baseline wall time number from
main, this PR fixes this.

Also, we make the CI fail if the time we take with the PRs code is more
than `TOLERANCE_PERCENTAGE` % slower, where for now
`TOLERANCE_PERCENTAGE=5.0%` (should be tuned based on the standard
deviation we observe in the next few PRs).

Last but not least, we stop mapping container ports to the host. It's
useless, and can create conflicts if you have other services running on
the same machine.