Lsst schema v9.0 (#233)
LSST has released the schema which is most likely to be what we've get
for the first real alerts put out by the survey. This PR:

- updates the structs to match the new schema. For that we mostly needed
to remove some fields that aren't in the new version, and add fields
that have been added. I did not add everything but only those that
seemed like they could be relevant and worth saving to the DB (which is
what we already do, we have a lot less fields in there then the full
schema)
- make the URL of the schema registry configurable. For a couple of
hours the new schema wasn't in the registry so I made my own mock
registry locally and being able to change the URL was a nice addition.
- replace the existing test LSST alert to have one that uses the new
schema, and update all LSST tests (since the values from the alert have
changed).
- commented out sections of the code that look for elements of the
schema that do not exist in version 9.0 (but did in version 7.1), and we
know will be re-implemented later.