Enrichment workers (#234)
Based on 7+ years of brokering ZTF alerts, it's very clear that there is
a number of features that most users compute themselves out of the alert
data. Looking at the filters in Kowalski, there is large number of
quantities that almost all filters rely on. Most of these are boolean
features that are computed from the `candidate` or from lightcurves.
While these are easy to compute in theory, things can get a little
convoluted when you try to implement them in a mongodb aggregation
pipeline (which is how our filters and Kowalski's work). Moreover, some
of Kowalski's users simply decided to not rely on it's filtering system
but on their own pipeline, since computing the features they needed was
just way too hard in Mongo.

So, here we:
- renamed the MLWorker EnrichmentWorker, where we now also compute a number of numerical features that are not ML-based.
- added `EnrichmentWorker`s for both LSST and DECAM, not just ZTF
- made a good number of changes to LSST and ZTF filter workers to
shorten the code by adding helper functions, and by "syncing" the LSST
filtering which was lagging being ZTF's (a number of changes made to ZTF
processing were not applied to LSST's). This meant fixing some unit
tests
- made forced photometry available in ZTF and LSST filtering, as well as
in the filtering output.