Throughput CI: rust build caching (in docker) (#242)
In this PR, we edit the `Dockerfile-benchmark` docker file to:
- first create an empty application with the same layout as booms, and
copy our Cargo.toml files to it
- then we build this empty project. This will download and build
dependencies. This way, this step only runs when there are Cargo.toml
changes, and otherwise is cached along with the build dependencies
- then we copy the source code of boom and build that. Since
dependencies are already built, this will only require build time for
the app itself. So, now that is the only build time that needs to happen
when there are source code changes, and we don't need to rerun the whole
dependencies update.

This is in an effort to cut down the docker build time for the
throughput test.

The throughput script wasn't waiting for the filter workers to
finish running. This PR also fixes that bug

Resolves #238