LSST Alert worker: survey xmatches (#245)
This PRs main goal is to get incoming LSST alerts xmatched with ZTF and
DECAM objects, like we already do for incoming ZTF alerts. We add a unit
test to make sure that LSST alerts are xmatched with ZTF as expected
(happy to add DECAM too, but not sure that would be of much use since
the exact same function is called for both survey anyways).

Also, I refactored some of the functions in the `filter` module to make
them a bit more generic, and modified LSST and ZTF's `build_alert`
functions to take them out of the trait so they can more easily be
re-used in other modules (for example, for the development of babamul).
I also made these functions accept batches of alerts rather than one
alert at a time, to reduce the number of queries we run to MongoDB
(instead of running N queries, where N is the number of alerts, we now
just run 1 query for a batch of alerts).