LSST alert: Missing IDs deserialization bugfix (#253)
Fix issue with lsst alerts giving us ids=0 (diaObjectId and ssObjectId, only one of these should be defined and the other None) instead of None. So, we simply deserialize 0 values as Nones when reading these ids.

This created a problem because only one of diaObjectId or ssObjectId should be defined values, but since the missing one was 0 we ran into our custom `AmbiguousIds` issue.

So, we remove the distinct deserialize objid vs ssoid functions to combine them into one function that can be used for both, and shorten it a bit.