API: Pipeline Query endpoint (#248)
This PR addresses part of #166, by adding a `post_pipeline_query`
endpoint, that lets a user of the API run a MongoDB aggregation pipeline
on a collection.

Also, this PR reorganizes the `queries.rs` file into a `queries` module
where we have a file per "category" of endpoint, to keep each feature a
little more self-contained and the code readable.

---------

Co-authored-by: Pete Bachant <pbachant@caltech.edu>