Filter worker: compress avro output (snappy) (#252)
This PR adds the `snappy` feature to apache-avro crate, so we can use
the snappy codec when making the avro packets that the filter worker
outputs to kafka. The avro packets we put out with the alert worker
contain images and lots of metadata, so adding compression on top is
very very appreciated. This means less data volume sent to Kafka and
fetched by clients, helping with networking.