Kafka timestamp based offsets - fix unit issue (#266)
I noticed that for LSST, pulling data from a specific date was still
giving me messages from before. Turns out the timestamps should be given
to kafka in milliseconds, not seconds. It doesn't make or break the
kafka pulling (as in we can still get messages, it just thinks we give
it an insanely old date so it starts from the beginning of the topic),
but we definitely want this bugfix merged before moving to prod.