General Dependencies Upgrade & Cleanup (2025/10/14) (#268)
We're getting closer and closer to deploying something in production,
and I think we should make an effort to keep up with the changes made to
our dependencies as much as we can.

This PR is just me running a `cargo upgrade` + `cargo upgrade
--incompatible` (that upgrades dependencies with potential breaking
changes that could prevent compiling our code).

NOTES:
- I didn't update `jsonwebtoken` to latest because it makes some major
changes, that as some users pointed out
https://github.com/Keats/jsonwebtoken/issues/449) adds a performance
hit. On my end I noticed that I now had to add the `Clone` trait to
`Claims`, which means that things that weren't cloned earlier now are
(and perhaps that contributes to the performance hit?).
- I removed `rsgen-avro` as a dev dependency. We don't need it in there,
it needs to be installed globally with `cargo install` (as described in
the README) and not as a code dependency. + by some convoluted
witchcraft more recent versions break `mongodb::bson::to_document()`'s
ability to read numerical values correctly???? It's all very odd, but
since we don't need it it's just not worth the headache. Additionally,
since that package needs an older version of apache avro that the one we
use/pin, it compiles it's own version just for that crate which means we
spend more time compiling crap we don't need.
- I also removed `criterion` as a dev dependency. We don't use it
anywhere as far as I can tell.
- I removed `flate2` since we use `zune-inflate` instead (have been for
some time now, it solved some issues I had with flate2)
- and finally, I removed `chrono` from the api crate since we only used
it in one line, that could easily be replaced by a call to `flare`'s
Time struct (which uses chrono internally, and that way we don't have 2
different versions compiled in our code). I reorganized the package list
alphabetically too, to keep things clean.