Kafka Consumers: rely on the group coordinator for consumer partition assignment (#269)
Instead of querying the kafka server for partition information in order to manually assign them to consumer processes locally, we now use a regular subscribe to rely on the server's group coordinator to assign partitions to our processes. This is coupled with an "on assign" callback (simply triggered by the first message reception, which happens after assignment), to then set the assigned partition's offsets to the required timestamp.