Missing ML docker volumes for the scheduler(s) (#284)
The models are not added in the Docker image nor made available through
volumes. We weill have multiple - and larger - models in the future and
want to use features to run them conditionally, so bringing them into
the container via volumes (rather than bloating the image with more
stuff) makes sense to me. If that becomes too inconvenient, we could
just put those in the image, at the cost of disk space usage.

Currently, this prevents the scheduler from working whatsoever in
Docker.