Multi-survey Filter workers Part 1: Filtering (#287)
* for each surveys' filter modules, add methods that can be used by
other surveys to bring in their data products (if there is a
match/alias).
* make use of these methods to let ZTF filters use LSST data and vice
versa
* when performing lookups to `aux` tables, we used to add 3 stages to
the pipelines:
* the `$lookup` itself, retrieving documents from the `aux` table by
`objectId`
* an `$addFields` to re-project the content of the lookup results (the
lookups gives you an array of matching documents but we know there is
only one match and we want to flatten these results to bring the match's
fields at the top level of the document, e.g. prv_candidates)
* Removed the `$unset` stage and instead set
the `aux` to `null` in the `$addFields` stage directly, saving us an
extra operation.