Filter: add name, description, and changelog fields (#298)
* update/replace some comments, add some warns for incorrect filters
that can't be loaded due to permissions
* add a 'name' field to the filters, so it can be a little easier for
users reading the output of boom or fetching filters to understand what
a filter is about
* add a 'description' field to the filters, again in an effort to have
better filter documentation. That one is optional.
* add a 'changelog' field to the filter versions, to give users a chance
to document changes from one version to another. It is optional and
always empty for the first version, since there are no changes with
regard to a previous version.
* update the add_filter cli to set names and fixed the permissions