Filter Test API: enhancement (#303)
* add sort and limit parameters to the Filter Test endpoint
* add a Filter Test Count endpoint, which returns a count rather than
documents. I made it a separate endpoint precisely because we don't need
any sort and limit parameters here, and because there is an opportunity
to skip some of the stages provided by the user since we just want to
get a count and not produce documents (i.e. we can skip some project
stages and such, basically anything that comes after the last match
stage of a pipeline).
* add a few utils like a SortOrder enum, and an response::ok_ser(x)
where x can be any struct that implements serde::Serialize rather than a
json value directly.
* define structs that represent the output of the filter test and filter
test count endpoints, to clearly document the output format in Utoipa +
return the pipeline that the server generated to retrieve
documents/count, alongside the actual result(s)