LSST alerts: compute JD (UTC) timestamp from MJD TAI (w/ hifitime) (#301)
use hifitime to compute a JD value from LSST's mjd tai (in utc, likefor
ZTF). Right now we ignored the fact that the MJD values from LSST aren't
in UTC but in TAI. So, the JD we computed by just adding the "MJD -> JD
offset" was also in JD. Issue is we rely on these timestamps to compare
times against ZTF, which is in UTC.

This PR fixes that by properly doing the TAI -> UTC conversion.

The start of exposure (ZTF) vs midpoint (LSST) issue remains, but we have no solution for it at the moment as both surveys do not systematically report exposure time along with their datapoints